[Refactor] 파일 위치 변경, 불필요한 파일 삭제
